Move `flight-cat` and `badges` roles to separate extension
