+++ orcestra_book/_ext/reports.py
@@ -30,49 +30,6 @@ def load_frontmatter(path, derive_flight=False):
     return frontmatter
 
 
-@lru_cache
-def create_flight_badge(src, cat_id):
-    """Return an HTML node based on a flight category id."""
-    with open(src / "reports" / "flight_categories.yaml", "r") as fp:
-        cat_tier = {
-            key: attrs["tier"]
-            for key, attrs in yaml.safe_load(fp)["categories"].items()
-        }.get(cat_id, "unknown")
-
-    span = f'<span class="badge cat-{cat_tier} cat-{cat_id}">{cat_id}</span>'
-    href = f'<a href="https://orcestra-campaign.org/search.html?q={cat_id}">{span}</a>'
-    node = nodes.raw(
-        text=href,
-        format="html",
-    )
-
-    return node
-
-
-class FlightCategoryRole(SphinxRole):
-    def run(self):
-        src = pathlib.Path(self.env.srcdir)
-        node = create_flight_badge(src, self.text)
-
-        return [node], []
-
-
-class BadgesRole(SphinxRole):
-    def run(self):
-        src = pathlib.Path(self.env.srcdir)
-        fm = load_frontmatter(self.env.doc2path(self.env.docname))
-        categories = fm.get("categories", [])
-
-        node_list = [create_flight_badge(src, cat_id) for cat_id in categories]
-
-        node = nodes.raw(
-            text=" ".join(n.astext() for n in node_list),
-            format="html",
-        )
-
-        return [node], []
-
-
 class LogoRole(SphinxRole):
     """Add a small campaign logo to the upper-right corner of a page."""
     def run(self):
@@ -137,8 +94,6 @@ def setup(app):
     app.connect("builder-inited", write_flight_table)
     app.connect("builder-inited", write_ship_table)
 
-    app.add_role("flight-cat", FlightCategoryRole())
-    app.add_role("badges", BadgesRole())
     app.add_role("logo", LogoRole())
 
     return {
